[ALPS08477052] cmdq: early porting for mt6991
1. early porting for mt6991
2. update mt6991 mminfra_bit
3. modify poll_timeout counter reg
4. get dbg value from tfa

MTK-Commit-Id: a6e7e8ddb0fd2c7e7a1cdc5896cda9e26cacf1d2

CR-Id: ALPS08477052
Feature: cmdq_gce

+++ drivers/misc/mediatek/mdp/cmdq_helper_ext.c
@@ -733,7 +733,7 @@ s32 cmdq_core_interpret_instruction(char *textBuf, s32 bufLen,
 						"Reg%d\n", arg_c_i);
 				} else {
 					reqLen = snprintf(textBuf, bufLen,
-						"%d\n", CMDQ_REG_REVERT_ADDR(
+						"%d\n", (u32)CMDQ_REG_REVERT_ADDR(
 						arg_c_i));
 				}
 				bufLen -= reqLen;
@@ -781,7 +781,7 @@ s32 cmdq_core_interpret_instruction(char *textBuf, s32 bufLen,
 					"Reg%d) jump ", arg_c_i);
 			} else {
 				reqLen = snprintf(textBuf, bufLen,
-					"%d) jump ", CMDQ_REG_REVERT_ADDR(
+					"%d) jump ", (u32)CMDQ_REG_REVERT_ADDR(
 					arg_c_i));
 			}
 			bufLen -= reqLen;
